# backend/internal/platform/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TeamOps API
  version: 1.0.0
  description: |
    REST API for multi-tenant project and task management.

    Access tokens are short-lived JWT bearer tokens. Refresh tokens are opaque,
    single-use credentials rotated by `/api/v1/auth/refresh`. Send JSON request
    bodies with `Content-Type: application/json`; unknown properties and trailing
    JSON values are rejected.

    Authorization is enforced server-side. OWNER has complete tenant control;
    ADMIN manages members, projects, tasks, labels, and audit logs but cannot
    delete the organization; MEMBER reads the tenant and creates/updates tasks
    and comments; VIEWER is read-only. Resource endpoints derive the organization
    from persisted projects/tasks to prevent IDOR.

    List endpoints use bounded offset pagination. `sort=-created_at` means
    descending; a `+` prefix means ascending. The optional `order=asc|desc`
    parameter is used when sort has no prefix. Project and task updates require
    the current `version`; stale versions return HTTP 409.

    Every `/api/v1` operation is subject to the configured IP rate limit and
    can return HTTP 429. JSON-writing operations can also return HTTP 413 or
    415, and unexpected server failures use the common error envelope with
    HTTP 500.
servers:
  - url: http://localhost:8080
    description: Local Docker/API server
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      operationId: register
      summary: Register a user and start a refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
            example: { email: ada@example.com, name: Ada Lovelace, password: 'Strong-Password-123!' }
      responses:
        '201': { description: User and token pair created, content: { application/json: { schema: { $ref: '#/components/schemas/TokenEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Authenticate with email and password
      description: Invalid email and invalid password deliberately return the same response to prevent user enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example: { email: ada@example.com, password: 'Strong-Password-123!' }
      responses:
        '200': { description: Authenticated token pair, content: { application/json: { schema: { $ref: '#/components/schemas/TokenEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refresh
      summary: Rotate a refresh token
      description: The supplied token is revoked atomically. Reuse revokes every active token in its family.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
            example: { refreshToken: 'base64url-opaque-refresh-token' }
      responses:
        '200': { description: Rotated token pair, content: { application/json: { schema: { $ref: '#/components/schemas/TokenEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Revoke a refresh-token family
      description: Idempotent for an unknown token; clients should discard both access and refresh tokens.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } } }
      responses:
        '204': { description: Session revoked }
        '400': { $ref: '#/components/responses/ValidationError' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/me:
    get:
      tags: [Users]
      operationId: currentUser
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Authenticated user, content: { application/json: { schema: { $ref: '#/components/schemas/UserEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      summary: List organizations for the current user
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }, { $ref: '#/components/parameters/OrganizationSort' }, { $ref: '#/components/parameters/SortOrder' }]
      responses:
        '200': { description: Organization page, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationListEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Create an organization; creator becomes OWNER
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationCreateRequest' }
            example: { name: Analytical Engines, slug: analytical-engines }
      responses:
        '201': { description: Organization created, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/organizations/{organizationId}:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
    get:
      tags: [Organizations]
      operationId: getOrganization
      summary: Read an organization as a member
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Organization, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      summary: Update organization settings (OWNER only)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationUpdateRequest' } } } }
      responses:
        '200': { description: Organization updated, content: { application/json: { schema: { $ref: '#/components/schemas/OrganizationEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [Organizations]
      operationId: deleteOrganization
      summary: Delete an organization (OWNER only)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Organization deleted }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/organizations/{organizationId}/members:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
    get:
      tags: [Organizations]
      operationId: listMembers
      summary: List organization members
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Memberships, content: { application/json: { schema: { $ref: '#/components/schemas/MembershipListEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Organizations]
      operationId: addOrUpdateMember
      summary: Add a member or update a non-owner role (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberUpsertRequest' }
            example: { email: grace@example.com, role: MEMBER }
      responses:
        '201': { description: Membership created or updated, content: { application/json: { schema: { $ref: '#/components/schemas/MembershipEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/organizations/{organizationId}/members/{userId}:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/UserId' }]
    delete:
      tags: [Organizations]
      operationId: removeMember
      summary: Remove a non-owner member (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Member removed }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/organizations/{organizationId}/projects:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List and filter projects as an organization member
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: archived, in: query, description: Exact archived-state filter, schema: { type: boolean } }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/ProjectSort' }
        - { $ref: '#/components/parameters/SortOrder' }
      responses:
        '200': { description: Project page, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectListEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreateRequest' }
            example: { name: Difference Engine, description: Delivery plan and milestones }
      responses:
        '201': { description: Project created, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/projects/{projectId}:
    parameters: [{ $ref: '#/components/parameters/ProjectId' }]
    get:
      tags: [Projects]
      operationId: getProject
      summary: Read a project as a member of its persisted organization
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Partially update a project (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectUpdateRequest' } } } }
      responses:
        '200': { description: Project updated, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Delete a project (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Project deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/projects/{projectId}/tasks:
    parameters: [{ $ref: '#/components/parameters/ProjectId' }]
    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List, filter, search, and sort tasks
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/TaskStatus' } }
        - { name: priority, in: query, schema: { $ref: '#/components/schemas/TaskPriority' } }
        - { name: assignee_id, in: query, description: Exact assignee UUID; `assigneeId` is also accepted, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/TaskSort' }
        - { $ref: '#/components/parameters/SortOrder' }
      responses:
        '200': { description: Task page, content: { application/json: { schema: { $ref: '#/components/schemas/TaskListEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a TODO task (OWNER, ADMIN, or MEMBER)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreateRequest' }
            example: { title: Draft algorithm notes, description: Prepare the first review, status: TODO, priority: HIGH, dueAt: '2026-10-15T12:00:00Z' }
      responses:
        '201': { description: Task created, content: { application/json: { schema: { $ref: '#/components/schemas/TaskEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/tasks/{taskId}:
    parameters: [{ $ref: '#/components/parameters/TaskId' }]
    get:
      tags: [Tasks]
      operationId: getTask
      summary: Read a task as a member of its persisted organization
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Task, content: { application/json: { schema: { $ref: '#/components/schemas/TaskEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      tags: [Tasks]
      operationId: updateTask
      summary: Update, assign, prioritize, schedule, or transition a task
      description: OWNER, ADMIN, or MEMBER. At least one mutable field and the expected current version are required. Explicit null clears assignee/due date. Invalid state transitions return 409.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskUpdateRequest' }
            example: { status: IN_PROGRESS, priority: URGENT, assigneeId: '11111111-1111-4111-8111-111111111111', dueAt: null, version: 3 }
      responses:
        '200': { description: Task updated and version incremented, content: { application/json: { schema: { $ref: '#/components/schemas/TaskEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: Delete a task (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Task deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/tasks/{taskId}/comments:
    parameters: [{ $ref: '#/components/parameters/TaskId' }]
    get:
      tags: [Comments]
      operationId: listTaskComments
      summary: List comments as an organization member
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Comments, content: { application/json: { schema: { $ref: '#/components/schemas/CommentListEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Comments]
      operationId: createTaskComment
      summary: Comment on a task (OWNER, ADMIN, or MEMBER)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CommentCreateRequest' } } } }
      responses:
        '201': { description: Comment created, content: { application/json: { schema: { $ref: '#/components/schemas/CommentEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/organizations/{organizationId}/labels:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
    get:
      tags: [Labels]
      operationId: listLabels
      summary: List organization labels
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Labels, content: { application/json: { schema: { $ref: '#/components/schemas/LabelListEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Labels]
      operationId: createLabel
      summary: Create an organization label (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LabelCreateRequest' } } } }
      responses:
        '201': { description: Label created, content: { application/json: { schema: { $ref: '#/components/schemas/LabelEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/tasks/{taskId}/labels:
    parameters: [{ $ref: '#/components/parameters/TaskId' }]
    put:
      tags: [Labels]
      operationId: replaceTaskLabels
      summary: Atomically replace task labels (OWNER, ADMIN, or MEMBER)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/SetLabelsRequest' } } } }
      responses:
        '204': { description: Labels replaced }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: 'Short-lived access token returned by register, login, or refresh.' }
...
  responses:
    ValidationError: { description: 'Malformed JSON, unknown field, invalid identifier, invalid filter, or business validation failure', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' }, example: { error: { code: validation_error, message: request validation failed, requestId: 8c69acaf-fda2-42bd-95c2-59a7ed3153d4 } } } } }
...
    Conflict: { description: 'Optimistic-lock version, uniqueness, immutable-owner, or state-transition conflict', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' }, example: { error: { code: version_conflict, message: resource was modified by another request, requestId: 8c69acaf-fda2-42bd-95c2-59a7ed3153d4 } } } } }
    BodyTooLarge: { description: Request body exceeds configured limit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
...
    RegisterRequest: { type: object, additionalProperties: false, required: [email, name, password], properties: { email: { type: string, format: email, maxLength: 254 }, name: { type: string, minLength: 2, maxLength: 100 }, password: { type: string, format: password, minLength: 12, maxLength: 72, description: 'Must contain upper/lowercase letters, a number, and a symbol.' } } }
...
    TokenEnvelope: { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/TokenResult' } } }
...